+++ client/package.json
@@ -51,6 +51,7 @@
     "@babel/core": "^7.24.7",
     "@babel/preset-env": "^7.24.7",
     "@babel/preset-react": "^7.24.7",
+    "@react-oauth/google": "^0.12.1",
     "babel-loader": "^9.1.3",
     "css-loader": "^7.1.2",
     "nodemon": "^3.1.3",
