+++ client/src/components/getstarted.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const GetStarted = () => {
+  return <div>Instructions Page</div>;
+};
+
+export default GetStarted;
