--- client/src/components/info.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Info = () => {
-  return (
-    <div>
-      <h2>Info Component</h2>
-      <p>This is the info page.</p>
-    </div>
-  );
-};
-
-export default Info;
