+++ client/src/index.js
@@ -4,6 +4,7 @@ import App from './App.js';
 import { configureStore } from '@reduxjs/toolkit';
 import globalReducer from './state';
 import { Provider } from 'react-redux';
+import { GoogleOAuthProvider } from '@react-oauth/google';
 // import { setupListeners } from '@reduxjs/toolkit/query';
 // import { api } from 'state/api.js';
 
@@ -23,7 +24,9 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <Provider store={store}>
-      <App />
+      <GoogleOAuthProvider clientId='1067747072451-8si6bdo5o4vk4arhvv74cc9m5ja7rcvf.apps.googleusercontent.com'>
+        <App />
+      </GoogleOAuthProvider>
     </Provider>
   </React.StrictMode>
 );
