+++ client/src/pages/signup.jsx
@@ -1,0 +1,181 @@
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import {
+  Container,
+  TextField,
+  Button,
+  Box,
+  Avatar,
+  Typography,
+  Alert,
+  Paper,
+} from '@mui/material';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import { useTheme } from '@mui/material/styles';
+
+const Signup = () => {
+  const [firstName, setFirstName] = useState('');
+  const [lastName, setLastName] = useState('');
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [error, setError] = useState('');
+  const [success, setSuccess] = useState('');
+  const navigate = useNavigate();
+  const theme = useTheme();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+
+    try {
+      const response = await fetch('http://localhost:3000/signup', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ username, password }),
+      });
+
+      const data = await response.json();
+      console.log(data);
+
+      if (response.ok) {
+        setSuccess('Signup successful!');
+        setTimeout(() => {
+          navigate('/credentials');
+        }, 2000);
+      } else {
+        setError(data.message);
+      }
+    } catch (error) {
+      console.error('Error:', error);
+      setError('An error occurred. Please try again.');
+    }
+  };
+
+  return (
+    <Container maxWidth='sm'>
+      <Box
+        sx={{
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+          marginTop: 8,
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component='h1' variant='h5'>
+          Sign Up
+        </Typography>
+        {error && (
+          <Alert severity='error' sx={{ mt: 2, width: '100%' }}>
+            {error}
+          </Alert>
+        )}
+        {success && (
+          <Alert severity='success' sx={{ mt: 2, width: '100%' }}>
+            {success}
+          </Alert>
+        )}
+        <Paper
+          elevation={3}
+          sx={{
+            mt: 1,
+            width: '100%',
+            padding: 4,
+            borderRadius: 2,
+            backgroundColor: theme.palette.primary.main,
+            color: theme.palette.primary.contrastText,
+          }}
+        >
+          <Box
+            component='form'
+            onSubmit={handleSubmit}
+            sx={{
+              display: 'flex',
+              flexDirection: 'column',
+              gap: 3,
+            }}
+          >
+            <TextField
+              variant='outlined'
+              label='First Name'
+              value={firstName}
+              onChange={(e) => setFirstName(e.target.value)}
+              required
+              fullWidth
+              InputLabelProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+              InputProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+            />
+            <TextField
+              variant='outlined'
+              label='Last Name'
+              value={lastName}
+              onChange={(e) => setLastName(e.target.value)}
+              required
+              fullWidth
+              InputLabelProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+              InputProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+            />
+            <TextField
+              variant='outlined'
+              label='Username'
+              value={username}
+              onChange={(e) => setUsername(e.target.value)}
+              required
+              fullWidth
+              InputLabelProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+              InputProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+            />
+            <TextField
+              variant='outlined'
+              label='Password'
+              type='password'
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+              fullWidth
+              InputLabelProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+              InputProps={{
+                style: { color: theme.palette.primary.contrastText },
+              }}
+            />
+            <Button
+              type='submit'
+              variant='contained'
+              color='secondary'
+              fullWidth
+              sx={{ padding: 1.5 }}
+            >
+              Sign Up
+            </Button>
+            <Button
+              onClick={() => navigate('/login')}
+              color='secondary'
+              sx={{ mt: 0.5 }}
+            >
+              Already have an account? Log in!
+            </Button>
+          </Box>
+        </Paper>
+      </Box>
+    </Container>
+  );
+};
+
+export default Signup;
